+++ changelog.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [unreleased]
 
+- Add NewSafePage and NewSafeHtmlEncoder with auto escaping of string values
 - Add attribute formaction
 
 ## [0.18.0] 2021-05-27